Add commands to react-server-cli
- start
- compile
- init (not yet implemented)
