+++ packages/react-server-cli/src/commands/init.js
@@ -1,0 +1,9 @@
+import callerDependency from "../callerDependency";
+
+const reactServer = require(callerDependency("react-server"));
+
+const logger = reactServer.logging.getLogger(__LOGGER__);
+
+export default function compile(){
+	logger.notice("Coming soon!");
+}
