+++ packages/react-server-test-pages/cli.js
@@ -1,5 +1,5 @@
 require("babel-core/register");
 
-const {start, parseCliArgs} = require("react-server-cli");
+const {run, parseCliArgs} = require("react-server-cli");
 
-start(parseCliArgs());
+run(parseCliArgs());
