+++ packages/react-server-test-pages/package.json
@@ -6,9 +6,9 @@
   "main": "index.js",
   "scripts": {
     "prepublish": "gulp build",
-    "start": "node ./cli.js",
+    "start": "node ./cli.js start",
     "test": "gulp test",
-    "debug": "node-debug --debug-brk=0 -p 9000 react-server-cli",
+    "debug": "node-debug --debug-brk=0 -p 9000 ./cli.js start",
     "clean": "rimraf npm-debug.log* __clientTemp target"
   },
   "author": "Bo Borgerson",
