+++ packages/react-server-website/.reactserverrc
@@ -1,19 +1,18 @@
 {
   "routes": "./routes.js",
   "host": "localhost",
-  "port": 3000,
-  "js-port": 3001,
+  "port": 3010,
+  "jsPort": 3011,
   "hot": false,
   "minify": false,
-  "long-term-caching": false,
-  "compile-only": false,
+  "longTermCaching": false,
   "https": false,
-  "log-level": "debug",
+  "logLevel": "debug",
   "env": {
-    "staging": {
-      "port": "4000"
-    },
     "production": {
+       "jsUrl": "/assets/",
+       "logLevel": "error",
+       "minify": true,
        "port": "80"
     }
   }
